providing a fix for aRMSD
It is not an elegant one, given the size of WinPython.exe
to get, and to extract.  But it is simplier to use this
-- even portable, on a USB thumb-drive, since PATH is not
affected --, and amend this by the VTK wheel, than plowing
through the dependendcies of aRMSD in Windows 7 (64 bit)
and PyInstaller.  It is limited in functionality (only
*.xyz model data are read) since openbabel is skipped here
intentionally, too, and some of the fancy renderings by
vtk (e.g., anaglyph representation) are equally skipped.

But as it is intended as a temporary fix, ok.